refactor(opencode): extract the session refresh reason classifier from the prompt delivery ledger
OpenCodePromptDeliveryLedger.ts mixed two concerns: the on-disk ledger
store and a self-contained text classifier that decides whether a reason
string or response state means "the OpenCode session needs refreshing".
The classifier is nine module-private helpers plus three regex tables
built around one public entry point, and it never touches the ledger
record shape.

Move that block verbatim into a new sibling module,
OpenCodeSessionRefreshReasonClassifier.ts. The three exported predicates
keep their names and signatures; the nine helpers and three patterns stay
module-private, exactly as private as they were before. The regexes,
including their /g flags and the sonarjs/regex-complexity disable comment,
are carried over character for character, so the lastIndex behaviour of
the two global patterns is unchanged - both are still used only with
.replace().

The three importers follow the symbols: the ledger itself now imports
isOpenCodeSessionRefreshResponseState for its two internal uses,
OpenCodePromptDeliveryFollowUpPolicy takes all three predicates from the
new module while keeping the retry constant and the two record types from
the ledger, and OpenCodeRuntimeDeliveryDiagnostics keeps only the ledger
record type, now as a plain type-only import in the trailing group.

No behaviour change. The ledger's existing test file passes untouched
apart from splitting one import line across the two modules; the 44
assertions in it and their test bodies are unchanged, which is the proof
that this is pure motion.

OpenCodePromptDeliveryLedger.ts: 1286 -> 1156 lines, still well above the
800-line floor and 130 under its frozen cap. The new module is 133 lines.
No baseline cap is lowered.
